pyproject.toml
[tool.poetry]
name = "saniraq-kz"
version = "0.1.0"
description = ""
authors = ["Zhangeldi <zhanikshakimurat@gmail.com>"]
readme = "README.md"
packages = [{include = "saniraq"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.11"
alembic = "^1.14.1"
sqlalchemy = "^2.0.38"
pydantic = "^2.10.6"
python-jose = "^3.4.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
psycopg2 = "^2.9.10"
bcrypt = "^4.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
